% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_db.R
\name{addDatabaseInternalColumns}
\alias{addDatabaseInternalColumns}
\title{Add Internal Columns to Database Tables}
\usage{
addDatabaseInternalColumns(tables)
}
\arguments{
\item{tables}{A list of data.table objects. Each data.table represents a table in the database
where internal columns need to be added.}
}
\value{
The modified list of data.tables with added internal columns.
}
\description{
This function modifies a list of data.tables by adding several internal columns
specific for database management. These columns include a unique ID column for each table,
columns for input and last check datetimes, and a column for the current dataset status.
All added columns are initialized with NA values. The unique ID column is placed as the
first column in each table.
}
\examples{
# Example usage:
library(data.table)
my_tables <- list(table1 = data.table(id = 1:3, data = c("A", "B", "C")),
                  table2 = data.table(id = 4:6, data = c("D", "E", "F")))
my_tables <- addDatabaseInternalColumns(my_tables)

# Now `my_tables` will have additional internal columns in each data.table

}
