% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_colourise_text.R
\name{cat_ok}
\alias{cat_ok}
\title{Print "OK" in Light Green Color}
\usage{
cat_ok(msg = "OK\\n", bg = NULL)
}
\arguments{
\item{msg}{The text message to be printed. Defaults to "OK".}

\item{bg}{The background color. Defaults to NULL.}
}
\description{
This function prints "OK" in light green color.
}
