% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_clock.R
\docType{class}
\name{Clock-class}
\alias{Clock-class}
\alias{Clock}
\title{Clock Class}
\description{
This class represents a clock with methods to manage and record time-related events.
This class is for logging nested processes with their run times and states. It
represents a clock that can be used to measure process times and maintain a
history of activities.
}

