% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_misc.R
\name{frame_string}
\alias{frame_string}
\title{Create a Framed String}
\usage{
frame_string(
  text =
    styled_string("\\nHello !!!\\n\\n\\nIs\\nthere\\n\\nA N Y O N E\\n\\nout\\nthere\\n???\\n "),
  pos = c("left", "center", "right")[1],
  edge = " ",
  hori = "-",
  vert = "|"
)
}
\arguments{
\item{text}{A character string to be framed.}

\item{pos}{The position of the framed text within the frame. It can be 'left', 'center', or 'right'.}

\item{edge}{The characters to be used for framing the top, bottom, left, and right edges.}

\item{hori}{The character to be used for horizontal framing.}

\item{vert}{The character to be used for vertical framing.}
}
\value{
A character string representing the framed text.
}
\description{
This function creates a framed string with specified formatting parameters.
}
\details{
The function creates a framed string by adding specified framing characters (edges, horizontal, and vertical) around the input text.
It allows customization of the frame's position, edge characters, and framing characters.
The resulting framed string is useful for creating visually appealing console outputs.
}
