% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_helper_functions.R
\name{getBeforeLastSlash}
\alias{getBeforeLastSlash}
\title{Extract Everything Before the Last Slash}
\usage{
getBeforeLastSlash(strings)
}
\arguments{
\item{strings}{A character vector where each element is a string potentially
containing slashes.}
}
\value{
A character vector of the same length as \code{strings}, where each element
is the part of the original string up to (but not including) the last slash.
}
\description{
This function takes a vector of strings and returns a new vector where each element
is modified to include only the portion of the original string that appears
before the last slash. The last slash itself is not included in the returned strings.
}
\examples{
absolute <- c('Patient/PID_001', 'Encounter/EID_001', 'Condition/CID_001', 'aaa/bbb/ccc', 'aaa')
getBeforeLastSlash(absolute)

}
