% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_db.R
\name{getDBConnection}
\alias{getDBConnection}
\title{Get a database connection}
\usage{
getDBConnection(user, password, dbname, host, port, schema)
}
\arguments{
\item{user}{The username for the database connection.}

\item{password}{The password for the database connection.}

\item{dbname}{The name of the database.}

\item{host}{The host address where the database is located.}

\item{port}{The port number for the database connection.}

\item{schema}{The schema under which the tables can be found}
}
\value{
A database connection object.
}
\description{
This function establishes a connection to a PostgreSQL database using the specified credentials.
}
