% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_helper_functions.R
\name{getGlobalVariablesByPrefix}
\alias{getGlobalVariablesByPrefix}
\title{Get a list of global variables with a specified prefix.}
\usage{
getGlobalVariablesByPrefix(prefix)
}
\arguments{
\item{prefix}{The prefix to match in variable names.}
}
\value{
A list containing the names and values of global variables with the given prefix.
}
\description{
This function searches for global variables in the current workspace whose names
start with the specified prefix and returns a list containing variable names
along with their values.
}
\examples{
\dontrun{
prefix <- "my_prefix"
result <- getGlobalVariablesByPrefix(prefix)
print(result)
}

}
\seealso{
\code{\link{ls}}, \code{\link{eapply}}
}
\keyword{global}
\keyword{prefix}
\keyword{variables}
\keyword{workspace}
