% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrputils-package.R
\docType{package}
\name{mrputils-package}
\alias{mrputils}
\alias{mrputils-package}
\title{mrputils: This are the specific utils for MRP detection}
\description{
This package was created for shared functions as utils in the MRP detection context (file, database, table, etc).
}
\author{
\strong{Maintainer}: Alexander Struebing \email{alexander.struebing@imise.uni-leipzig.de} (\href{https://orcid.org/0000-0001-5151-7665}{ORCID})

Authors:
\itemize{
  \item Florian Schmidt \email{florian.schmidt@imise.uni-leipzig.de} (\href{https://orcid.org/0000-0003-2027-8213}{ORCID})
}

}
\keyword{internal}
