% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_misc.R
\name{polar_disclaimer}
\alias{polar_disclaimer}
\title{Print a Disclaimer Text}
\usage{
polar_disclaimer()
}
\description{
This function prints a disclaimer text with detailed information about the project settings,
system details, hardware information, and request parameters.
}
\details{
The disclaimer text includes information about the project name, time stamp, R version, system details,
hardware information, working directories, request parameters, and software versions.
It provides a comprehensive overview of the project configuration.
}
