% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_sample_resources.R
\name{polar_get_resources_by_ids}
\alias{polar_get_resources_by_ids}
\title{Get FHIR Resources by IDs}
\usage{
polar_get_resources_by_ids(
  endpoint,
  resource,
  ids,
  id_param_str = "_id",
  parameters = polar_add_common_request_params(c()),
  verbose = 0
)
}
\arguments{
\item{endpoint}{The FHIR server endpoint URL.}

\item{resource}{The name of the FHIR resource to retrieve.}

\item{ids}{A vector of identifiers for the FHIR resources.}

\item{id_param_str}{The parameter string for specifying the resource IDs (default is '_id').}

\item{parameters}{Additional parameters for the FHIR resource query (optional).}

\item{verbose}{The level of verbosity for logging (default is 0).}
}
\value{
A list of FHIR resources in the form of a fhir_bundle_list.
}
\description{
This function retrieves FHIR resources from a server based on a list of resource IDs.
It supports both GET and POST methods for querying the server.
}
