% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_misc.R
\name{polar_run}
\alias{polar_run}
\title{Run and log a process}
\usage{
polar_run(
  message,
  process,
  single_line = TRUE,
  throw_exception = TRUE,
  verbose = VERBOSE
)
}
\arguments{
\item{message}{A character of length one. Has to be unique}

\item{process}{An expression}

\item{single_line}{A logical of length one}

\item{throw_exception}{A logical of length one. Should catched error throw again}

\item{verbose}{The verbose level}
}
\value{
The value process returns.
}
\description{
Run and log a process
}
