% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_dir.R
\name{polar_save_bundles}
\alias{polar_save_bundles}
\title{Save Bundle
bndls <- fhircrackr::fhir_search("http://vonk.fire.ly/R4/Patient?gender=male", max_bundles = 1)
polar_save_bundles(bndls)}
\usage{
polar_save_bundles(bundles)
}
\arguments{
\item{bundles}{A fhir_bundle_list containing the bundles to store to the sub project related bundles directory.}
}
\value{
Nothing
}
\description{
Save Bundle
bndls <- fhircrackr::fhir_search("http://vonk.fire.ly/R4/Patient?gender=male", max_bundles = 1)
polar_save_bundles(bndls)
}
