% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_dir.R
\name{polar_save_result_table_as_csv}
\alias{polar_save_result_table_as_csv}
\title{Save a Result Table of an \emph{\strong{ANALYSIS-SCTIPT}} in the \emph{public} \code{results} directory to which was created for the specific subproject.}
\usage{
polar_save_result_table_as_csv(
  table,
  filename_without_extension = NA,
  extension = ".txt",
  sep = "\\t",
  dec = ".",
  collapse = " ~ ",
  row.names = FALSE
)
}
\arguments{
\item{table}{the table to save}

\item{filename_without_extension}{If the default NA is not changed then the name is the name of the table variable.}

\item{extension}{the file name extension (default is ".txt")}

\item{sep}{the field separator string. Values within each row of x are separated by this string.}

\item{dec}{the string to use for decimal points in numeric or complex columns: must be a single character}

\item{collapse}{If there are any list columns then all list items will be pasted with this delimiter to one string
and the class of the column will be changed to character. Default is ' ~ '.}

\item{row.names}{either a logical value indicating whether the row names of x are to be written along with x, or a
character vector of row names to be written.}
}
\value{
Nothing.
}
\description{
Save a Result Table of an \emph{\strong{ANALYSIS-SCTIPT}} in the \emph{public} \code{results} directory to which was created for the specific subproject.
}
