% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_dir.R
\name{polar_write_rdata}
\alias{polar_write_rdata}
\title{Save an Object as RDS-File in the \emph{private} \code{tables} directory to which was created for the specific subproject.}
\usage{
polar_write_rdata(
  object = table,
  filename_without_extension = NA,
  project_sub_dir = NA
)
}
\arguments{
\item{object}{the table to write}

\item{filename_without_extension}{If the default NA is not changed then the name is the name of the table variable.}

\item{project_sub_dir}{subdirectory of the current working directory where the table is located. If NA (default),
then the tables will be loaded from outputLocal/tables'.}
}
\value{
Nothing.
}
\description{
a <- 1
polar_write_rdata(a)
}
