% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_helper_functions.R
\name{run_in}
\alias{run_in}
\title{Execute an inner script with a specified message and process}
\usage{
run_in(message, process)
}
\arguments{
\item{message}{A character string describing the purpose of the inner script.}

\item{process}{A function representing the inner script to be executed.}
}
\description{
This function runs an inner script with the provided message and process, controlling
the verbosity level.
}
