% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_db.R
\name{showTablesStructure}
\alias{showTablesStructure}
\title{Show the structure of tables in a database}
\usage{
showTablesStructure(db_connection)
}
\arguments{
\item{db_connection}{A valid database connection object.}
}
\description{
This function retrieves the list of existing table names in a database connection and displays their structure.
}
