% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_table.R
\name{toMatrix}
\alias{toMatrix}
\title{Convert a list into a matrix with specified column count and fill value.}
\usage{
toMatrix(list, colCount, fill = NA)
}
\arguments{
\item{list}{A list that should be converted into a matrix.}

\item{colCount}{The number of columns in the resulting matrix.}

\item{fill}{The value to fill missing values in the last row if the element count is not a
multiple of the column count (default is NA).}
}
\value{
A matrix with the specified column count and optional padding in the last row.
}
\description{
This function takes a list and converts it into a matrix with the specified column count.
If the number of elements in the list is not a multiple of the column count, the 'fill' value
is used to pad the last row of the matrix. The resulting matrix is created with 'colCount' columns
and the number of rows is determined by dividing the length of the list by 'colCount'.
}
\examples{
# Example 1: Convert a list into a matrix with 2 columns
my_list <- c(1, 2, 3, 4, 5, 6)
toMatrix(my_list, 2)

# Example 2: Convert a list into a matrix with 3 columns and use 0 as the fill value
another_list <- c(1, 2, 3, 4, 5, 6, 7)
toMatrix(another_list, 3, fill = 0)

# Example 3: Convert a list into a matrix with 4 columns and use "NA" as the fill value
yet_another_list <- c(1, 2, 3, 4, 5, 6, 7, 8)
toMatrix(yet_another_list, 4, fill = "NA")

}
