% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{VL_00_NOTHING}
\alias{VL_00_NOTHING}
\title{Verbose Level Constants}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
VL_00_NOTHING
}
\description{
This file defines constants for different verbose levels.
}
\details{
docType package
name constants
aliases constants
keywords internal

export not needed !?
}
\keyword{datasets}
